requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Flask-Login==0.6.2
Flask-Migrate==4.0.4
pymysql==1.1.0
email_validator==2.0.0
python-dotenv==1.0.0
Werkzeug==2.3.6
Flask-Bootstrap==3.3.7.1